+++ code-snippets/configuration/register_data_sources.rs
@@ -11,7 +11,6 @@
 //! - Registering data sources with register_data_source()
 //! - Setting default configuration
 
-use sz_rust_sdk::helpers::ExampleEnvironment;
 use sz_rust_sdk::prelude::*;
 
 fn main() -> SzResult<()> {