+++ code-snippets/searching/search_records.rs
@@ -10,7 +10,6 @@
 //! - Using search flags for different result formats
 
 use serde_json::{json, Value};
-use sz_rust_sdk::helpers::ExampleEnvironment;
 use sz_rust_sdk::prelude::*;
 
 fn main() -> SzResult<()> {